increase index size for grids to accommodate larger problems
Grid indexes are not especially performance sensitive, and we're
sticking it in a machine int anyway.

+++ grid.hh
@@ -11,7 +11,7 @@ using namespace std;
 struct Grid {
   struct Node {
     typedef uint8_t ordinate_t;
-    typedef uint8_t index_t;
+    typedef unsigned int index_t;
     typedef int8_t degree_t; // may become negative
     typedef pair<ordinate_t, ordinate_t> coordinate_t;
